Move healthQuestionsForDecision method to parent Reply class
