+++ app/models/consent.js
@@ -1,81 +1,13 @@
-import vaccines from '../datasets/vaccines.js'
 import { hasAnswersNeedingTriage } from '../utils/reply.js'
 import { formatLinkWithSecondaryText } from '../utils/string.js'
 
-import { ProgrammeType } from './programme.js'
-import { Reply, ReplyDecision } from './reply.js'
-import { VaccineMethod } from './vaccine.js'
+import { Reply } from './reply.js'
 
 /**
  * @class Consent
  * @augments Reply
  */
 export class Consent extends Reply {
-  /**
-   * Get health questions to show based on programme and decision given
-   *
-   * @returns {Map} - Health questions
-   */
-  get healthQuestionsForDecision() {
-    const { Flu, HPV, MenACWY, TdIPV } = ProgrammeType
-    const { Injection, Nasal } = VaccineMethod
-    const programme = this.session.primaryProgrammes[0]
-
-    const healthQuestionsForDecision = new Map()
-    let consentedMethod
-    let consentedVaccine
-
-    // Consent given for flu programme with method of vaccination
-    if (programme.type === Flu) {
-      consentedMethod =
-        this.decision === ReplyDecision.OnlyFluInjection ? Injection : Nasal
-      consentedVaccine = Object.values(vaccines).find(
-        (programme) =>
-          programme.type === Flu && programme.method === consentedMethod
-      )
-    }
-
-    // Consent given for HPV programme
-    if (programme.type === HPV) {
-      consentedVaccine = Object.values(vaccines).find(
-        (programme) => programme.type === HPV
-      )
-    }
-
-    // Consent given for MenACWY programme only
-    if (this.decision === ReplyDecision.OnlyMenACWY) {
-      consentedVaccine = Object.values(vaccines).find(
-        (programme) => programme.type === MenACWY
-      )
-    }
-
-    // Consent given for Td/IPV programme only
-    if (this.decision === ReplyDecision.OnlyTdIPV) {
-      consentedVaccine = Object.values(vaccines).find(
-        (programme) => programme.type === TdIPV
-      )
-    }
-
-    // Consent given for all programmes
-    if (ReplyDecision.Given && !consentedVaccine) {
-      consentedVaccine = this.session.vaccines
-    }
-
-    const consentedVaccines = Array.isArray(consentedVaccine)
-      ? consentedVaccine
-      : [consentedVaccine]
-    for (const vaccine of consentedVaccines) {
-      for (const [key, value] of Object.entries(vaccine.healthQuestions)) {
-        healthQuestionsForDecision.set(key, value)
-      }
-    }
-
-    // Always ask support question last
-    healthQuestionsForDecision.set('support', {})
-
-    return healthQuestionsForDecision
-  }
-
   /**
    * Answers in this consent response need triage
    *
